fixing summary counts. Removed the interface that is no longer needed.

+++ api/src/main/java/ca/bc/gov/educ/studentdatacollection/api/struct/v1/SdcSchoolCollectionStudentValidationIssueErrorWarningCount.java
@@ -19,8 +19,8 @@ public class SdcSchoolCollectionStudentValidationIssueErrorWarningCount implemen
   @Serial
   private static final long serialVersionUID = 1L;
 
-  private Integer error = 0;
+  private long error = 0;
 
-  private Integer warning = 0;
+  private long warning = 0;
 
 }
